+++ linbox/algorithms/gauss/gauss.inl
@@ -4,20 +4,20 @@
  * Written by Jean-Guillaume Dumas <Jean-Guillaume.Dumas@imag.fr>
  * Time-stamp: <18 Jun 10 15:48:38 Jean-Guillaume.Dumas@imag.fr>
  *
- * 
+ *
  * ========LICENCE========
  * This file is part of the library LinBox.
- * 
+ *
  * LinBox is free software: you can redistribute it and/or modify
  * it under the terms of the  GNU Lesser General Public
  * License as published by the Free Software Foundation; either
  * version 2.1 of the License, or (at your option) any later version.
- * 
+ *
  * This library is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  * Lesser General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
@@ -92,8 +92,8 @@ namespace LinBox
 
 		// assignment of LigneA with the domain object
 		for (unsigned long jj = 0; jj < Ni; ++jj)
-			for (unsigned long k = 0; k < LigneA[jj].size (); k++)
-				++col_density[LigneA[jj][k].first];
+			for (unsigned long k = 0; k < LigneA[(size_t)jj].size (); k++)
+				++col_density[LigneA[(size_t)jj][k].first];
 
 		long last = (long)Ni - 1;
 		long c;
@@ -121,7 +121,7 @@ namespace LinBox
 #ifdef __LINBOX_FILLIN__
 				long sl(0);
 				for (size_t l = 0; l < Ni; ++l)
-					sl += LigneA[l].size ();
+					sl +=(long) LigneA[(size_t)l].size ();
 
 				commentator().report (Commentator::LEVEL_IMPORTANT, PARTIAL_RESULT)
 				<< "Fillin (" << Rank << "/" << Ni << ") = "
@@ -134,7 +134,7 @@ namespace LinBox
 
 			long l;
 			for(l = k; l < static_cast<long>(Ni); ++l) {
-				if ( (s = LigneA[l].size()) ) {
+				if ( (s = (long) LigneA[(size_t)l].size()) ) {
 					p = l;
 					break;
 				}
@@ -144,7 +144,7 @@ namespace LinBox
 				long sl;
 				// Row permutation for the sparsest row
 				for (; l < static_cast<long>(Ni); ++l)
-					if (((sl = LigneA[l].size ()) < s) && (sl)) {
+					if (((sl =(long) LigneA[(size_t)l].size ()) < s) && (sl)) {
 						s = sl;
 						p = l;
 					}
@@ -153,8 +153,8 @@ namespace LinBox
 					// std::cerr << "Permuting rows: " << k << " <--> " << p << std::endl;
 					invQ.push_front( std::pair<size_t,size_t>((size_t)k,(size_t)p) );
 					field().negin(determinant);
-					std::swap( *LigneA_k, LigneA[p]);
-					std::swap( LigneL[k], LigneL[p]);
+					std::swap( *LigneA_k, LigneA[(size_t)p]);
+					std::swap( LigneL[(size_t)k], LigneL[(size_t)p]);
 				}
 
 
@@ -163,16 +163,16 @@ namespace LinBox
 				if (c != -1) {
 					long ll;
 					if ( c != (static_cast<long>(Rank)-1) ) {
-						P.permute(Rank-1,c);
+						P.permute(Rank-1,(size_t)c);
 						for (ll=0      ; ll < k ; ++ll)
-							permute( LigneA[ll], Rank, c);
+							permute( LigneA[(size_t)ll], Rank, c);
 					}
-					long npiv=LigneA_k->size();
+					long npiv=(long)LigneA_k->size();
 					for (ll = k+1; ll < static_cast<long>(Ni); ++ll) {
 						E hc;
 						hc.first=(unsigned)Rank-1;
-						eliminate (hc.second, LigneA[ll], *LigneA_k, Rank, c, npiv, col_density);
-						if(! field().isZero(hc.second)) LigneL[ll].push_back(hc);
+						eliminate (hc.second, LigneA[(size_t)ll], *LigneA_k, Rank, c, (size_t)npiv, col_density);
+						if(! field().isZero(hc.second)) LigneL[(size_t)ll].push_back(hc);
 					}
 				}
 
@@ -182,25 +182,25 @@ namespace LinBox
 #endif
 			}
 			E one((unsigned)k,Eone);
-			LigneL[k].push_back(one);
+			LigneL[(size_t)k].push_back(one);
 			//                 LigneL.write(rep << "L:= ", FORMAT_MAPLE) << std::endl;
 			//                 LigneA.write(rep << "U:= ", FORMAT_MAPLE) << std::endl;
 		}//for k
 
-		SparseFindPivot ( LigneA[last], Rank, c, determinant);
+		SparseFindPivot ( LigneA[(size_t)last], Rank, c, determinant);
 		if (c != -1) {
 			if ( c != (static_cast<long>(Rank)-1) ) {
-				P.permute(Rank-1,c);
+				P.permute(Rank-1,(size_t)c);
 				for (long ll=0      ; ll < last ; ++ll)
-					permute( LigneA[ll], Rank, c);
+					permute( LigneA[(size_t)ll], Rank, c);
 			}
 		}
 
 		E one((unsigned)last,Eone);
-		LigneL[last].push_back(one);
+		LigneL[(size_t)last].push_back(one);
 
 #ifdef __LINBOX_COUNT__
-		nbelem += LigneA[last].size ();
+		nbelem += LigneA[(size_t)last].size ();
 		commentator().report (Commentator::LEVEL_NORMAL, PARTIAL_RESULT)
 		<< "Left elements : " << nbelem << std::endl;
 #endif
@@ -208,7 +208,7 @@ namespace LinBox
 #ifdef __LINBOX_FILLIN__
 		long sl(0);
 		for (size_t l=0; l < Ni; ++l)
-			sl += LigneA[l].size ();
+			sl += (long) LigneA[(size_t)l].size ();
 
 		commentator().report (Commentator::LEVEL_IMPORTANT, PARTIAL_RESULT)
 		<< "Fillin (" << Rank << "/" << Ni << ") = " << sl
@@ -274,8 +274,8 @@ namespace LinBox
 
 		// assignment of LigneA with the domain object
 		for (unsigned long jj = 0; jj < Ni; ++jj)
-			for (unsigned long k = 0; k < LigneA[jj].size (); k++)
-				++col_density[LigneA[jj][k].first];
+			for (unsigned long k = 0; k < LigneA[(size_t)jj].size (); k++)
+				++col_density[LigneA[(size_t)jj][k].first];
 
 		long last = (long)Ni - 1;
 		long c;
@@ -290,7 +290,7 @@ namespace LinBox
 		// Elimination steps with reordering
 		for (long k = 0; k < last; ++k) {
 			unsigned long l;
-			long p = k, s = LigneA[k].size (), sl;
+			long p = k, s = (long)LigneA[(size_t)k].size (), sl;
 
 #ifdef __LINBOX_FILLIN__
 			if ( ! (k % 100) ) {
@@ -300,7 +300,7 @@ namespace LinBox
 					commentator().progress (k);
 #ifdef __LINBOX_FILLIN__
 					for (sl = 0, l = 0; l < Ni; ++l)
-						sl += LigneA[l].size ();
+						sl += (long)LigneA[(size_t)l].size ();
 
 					commentator().report (Commentator::LEVEL_IMPORTANT, PARTIAL_RESULT)
 					<< "Fillin (" << Rank << "/" << Ni << ") = "
@@ -314,40 +314,40 @@ namespace LinBox
 				if (s) {
 					// Row permutation for the sparsest row
 					for (l = (unsigned long)k + 1; l < (unsigned long)Ni; ++l)
-						if (((sl = LigneA[l].size ()) < s) && (sl)) {
+						if (((sl = (long)LigneA[(size_t)l].size ()) < s) && (sl)) {
 							s = sl;
 							p = (long)l;
 						}
 
 					if (p != k) {
 						field().negin(determinant);
-						Vector vtm = LigneA[k];
-						LigneA[k] = LigneA[p];
-						LigneA[p] = vtm;
+						Vector vtm = LigneA[(size_t)k];
+						LigneA[(size_t)k] = LigneA[(size_t)p];
+						LigneA[(size_t)p] = vtm;
 					}
 
 					//                     LigneA.write(std::cerr << "BEF, k:" << k << ", Rank:" << Rank << ", c:" << c)<<std::endl;
 
-					SparseFindPivot (LigneA[k], Rank, c, col_density, determinant);
+					SparseFindPivot (LigneA[(size_t)k], Rank, c, col_density, determinant);
 					//                     LigneA.write(std::cerr << "PIV, k:" << k << ", Rank:" << Rank << ", c:" << c)<<std::endl;
 					if (c != -1) {
 						for (l = (unsigned long)k + 1; l < (unsigned long)Ni; ++l)
-							eliminate (LigneA[l], LigneA[k], Rank, c, col_density);
+							eliminate (LigneA[(size_t)l], LigneA[(size_t)k], Rank, c, col_density);
 					}
 
 					//                     LigneA.write(std::cerr << "AFT " )<<std::endl;
 #ifdef __LINBOX_COUNT__
-					nbelem += LigneA[k].size ();
+					nbelem += LigneA[(size_t)k].size ();
 #endif
-					LigneA[k] = Vzer;
+					LigneA[(size_t)k] = Vzer;
 				}
 
 			}//for k
 
-			SparseFindPivot (LigneA[last], Rank, c, determinant);
+			SparseFindPivot (LigneA[(size_t)last], Rank, c, determinant);
 
 #ifdef __LINBOX_COUNT__
-			nbelem += LigneA[last].size ();
+			nbelem += LigneA[(size_t)last].size ();
 			commentator().report (Commentator::LEVEL_NORMAL, PARTIAL_RESULT)
 			<< "Left elements : " << nbelem << std::endl;
 #endif
@@ -355,7 +355,7 @@ namespace LinBox
 #ifdef __LINBOX_FILLIN__
 			long sl(0);
 			for (size_t l=0; l < Ni; ++l)
-				sl += LigneA[l].size ();
+				sl += (long)LigneA[(size_t)l].size ();
 
 			commentator().report (Commentator::LEVEL_IMPORTANT, PARTIAL_RESULT)
 			<< "Fillin (" << Rank << "/" << Ni << ") = " << sl
@@ -409,8 +409,8 @@ namespace LinBox
 
 			// assignment of LigneA with the domain object
 			for (unsigned long jj = 0; jj < Ni; ++jj)
-				for (unsigned long k = 0; k < LigneA[jj].size (); k++)
-					++col_density[LigneA[jj][k].first];
+				for (unsigned long k = 0; k < LigneA[(size_t)jj].size (); k++)
+					++col_density[LigneA[(size_t)jj][k].first];
 
 			long last = (long)Ni - 1;
 			long c;
@@ -425,7 +425,7 @@ namespace LinBox
 			// Elimination steps with reordering
 			for (long k = 0; k < last; ++k) {
 				unsigned long l;
-				long p = k, s = LigneA[k].size (), sl;
+				long p = k, s =(long) LigneA[(size_t)k].size (), sl;
 
 #ifdef __LINBOX_FILLIN__
 				if ( ! (k % 100) )
@@ -436,7 +436,7 @@ namespace LinBox
 					commentator().progress (k);
 #ifdef __LINBOX_FILLIN__
 					for (sl = 0, l = 0; l < Ni; ++l)
-						sl += LigneA[l].size ();
+						sl +=(long) LigneA[(size_t)l].size ();
 
 					commentator().report (Commentator::LEVEL_IMPORTANT, PARTIAL_RESULT)
 					<< "Fillin (" << Rank << "/" << Ni << ") = "
@@ -450,50 +450,50 @@ namespace LinBox
 				if (s) {
 					// Row permutation for the sparsest row
 					for (l = (unsigned long)k + 1; l < (unsigned long)Ni; ++l)
-						if (((sl = LigneA[l].size ()) < s) && (sl)) {
+						if (((sl =(long) LigneA[(size_t)l].size ()) < s) && (sl)) {
 							s = sl;
 							p = (long)l;
 						}
 
 					if (p != k) {
 						field().negin(determinant);
-						Vector vtm = LigneA[k];
-						LigneA[k] = LigneA[p];
-						LigneA[p] = vtm;
+						Vector vtm = LigneA[(size_t)k];
+						LigneA[(size_t)k] = LigneA[(size_t)p];
+						LigneA[(size_t)p] = vtm;
 					}
 
 					//                     LigneA.write(std::cerr << "BEF, k:" << k << ", Rank:" << Rank << ", c:" << c)<<std::endl;
 
-					SparseFindPivot (LigneA[k], Rank, c, col_density, determinant);
+					SparseFindPivot (LigneA[(size_t)k], Rank, c, col_density, determinant);
 					//                     LigneA.write(std::cerr << "PIV, k:" << k << ", Rank:" << Rank << ", c:" << c)<<std::endl;
 					if (c != -1) {
 						if ( c != (static_cast<long>(Rank)-1) )
-							P.permute(Rank-1,c);
+							P.permute(Rank-1,(size_t)c);
 						for (long ll=0; ll < k ; ++ll)
-							permute( LigneA[ll], Rank, c);
+							permute( LigneA[(size_t)ll], Rank, c);
 
 						for (l = (unsigned long)k + 1; l < (unsigned long)Ni; ++l)
-							eliminate (LigneA[l], LigneA[k], Rank, c, col_density);
+							eliminate (LigneA[(size_t)l], LigneA[(size_t)k], Rank, c, col_density);
 					}
 
 					//                     LigneA.write(std::cerr << "AFT " )<<std::endl;
 #ifdef __LINBOX_COUNT__
-					nbelem += LigneA[k].size ();
+					nbelem += LigneA[(size_t)k].size ();
 #endif
 				}
 
 			}//for k
 
-			SparseFindPivot (LigneA[last], Rank, c, determinant);
+			SparseFindPivot (LigneA[(size_t)last], Rank, c, determinant);
 			if ( (c != -1) && (c != (static_cast<long>(Rank)-1) ) ) {
-				P.permute(Rank-1,c);
+				P.permute(Rank-1,(size_t)c);
 				for (long ll=0; ll < last ; ++ll)
-					permute( LigneA[ll], Rank, c);
+					permute( LigneA[(size_t)ll], Rank, c);
 			}
 
 
 #ifdef __LINBOX_COUNT__
-			nbelem += LigneA[last].size ();
+			nbelem += LigneA[(size_t)last].size ();
 			commentator().report (Commentator::LEVEL_NORMAL, PARTIAL_RESULT)
 			<< "Left elements : " << nbelem << std::endl;
 #endif
@@ -501,7 +501,7 @@ namespace LinBox
 #ifdef __LINBOX_FILLIN__
 			long sl(0);
 			for (size_t l=0; l < Ni; ++l)
-				sl += LigneA[l].size ();
+				sl +=(long) LigneA[(size_t)l].size ();
 
 			commentator().report (Commentator::LEVEL_IMPORTANT, PARTIAL_RESULT)
 			<< "Fillin (" << Rank << "/" << Ni << ") = " << sl
@@ -562,23 +562,23 @@ namespace LinBox
 
 				unsigned long l;
 
-				if (!LigneA[k].empty ()) {
-					SparseFindPivot (LigneA[k], indcol, c, determinant);
+				if (!LigneA[(size_t)k].empty ()) {
+					SparseFindPivot (LigneA[(size_t)k], indcol, c, determinant);
 					if (c !=  -1)
 						for (l = (unsigned long)k + 1; l < (unsigned long)Ni; ++l)
-							eliminate (LigneA[l], LigneA[k], indcol, c);
+							eliminate (LigneA[(size_t)l], LigneA[(size_t)k], indcol, c);
 
 #ifdef __LINBOX_COUNT__
-					nbelem += LigneA[k].size ();
+					nbelem += LigneA[(size_t)k].size ();
 #endif
-					LigneA[k] = Vzer;
+					LigneA[(size_t)k] = Vzer;
 				}
 			}
 
-			SparseFindPivot ( LigneA[last], indcol, c, determinant);
+			SparseFindPivot ( LigneA[(size_t)last], indcol, c, determinant);
 
 #ifdef __LINBOX_COUNT__
-			nbelem += LigneA[last].size ();
+			nbelem += LigneA[(size_t)last].size ();
 			commentator().report (Commentator::LEVEL_NORMAL, PARTIAL_RESULT)
 			<< "Left elements : " << nbelem << std::endl;
 #endif
